+++ design-system/eslint-plugin/CHANGELOG.md
@@ -1,5 +1,46 @@
 # @atlaskit/eslint-plugin-design-system
 
+## 0.0.4
+
+### Patch Changes
+
+- [`1926dba3536`](https://bitbucket.org/atlassian/atlassian-frontend/commits/1926dba3536) - Adds, removes & renames tokens
+
+  Adds:
+
+  - `color.backgroundSelect`
+
+  Renames:
+
+  - `color.borderTextHighlighted` to `color.bordertextSelected`
+  - `elevation.base` to `evelation.backgroundDefault`
+  - `elevation.flatSecondary` to `elevation.backgroundSunken`
+  - `elevation.backgroundCard` to `color.backgroundCard`
+  - `elevation.backgroundOverlay` to `color.backgroundOverlay`
+  - `elevation.borderOverlay` to `color.borderOverlay`
+  - `elevation.shadowCard` to `shadow.card`
+  - `elevation.shadowOverlay` to `shadow.overlay`
+
+  Removes:
+
+  - `elevation.boarderFlatPrimary`
+
+  Updates:
+
+  - `elevation.shadowOverlay` value to `DN100`
+  - `color.textWarning` in light mode to `O800`
+  - `color.iconBorderWarning` in light mode to `O600`
+
+- Updated dependencies
+
+## 0.0.3
+
+### Patch Changes
+
+- [`ade8d954aa5`](https://bitbucket.org/atlassian/atlassian-frontend/commits/ade8d954aa5) - Out of the box configs have been removed until stable release.
+- [`f2a0a48903d`](https://bitbucket.org/atlassian/atlassian-frontend/commits/f2a0a48903d) - Errors no longer show up on import declarations.
+- [`b71d3cd3d2f`](https://bitbucket.org/atlassian/atlassian-frontend/commits/b71d3cd3d2f) - Internal artefacts no longer make their way to npm.
+
 ## 0.0.2
 
 ### Patch Changes